backend/requirements.txt
flask==2.3.2
flask-sqlalchemy==3.0.3
flask-migrate==4.0.4
flask-cors==4.0.0
flask-jwt-extended==4.5.2
python-dotenv==1.0.0
werkzeug==2.3.4
sqlalchemy==2.0.15
psycopg2-binary==2.9.6
python-dateutil==2.8.2
alembic==1.10.4
gunicorn==20.1.0
